third_party/blink/web_tests/external/wpt/wasm/core/gc/br_on_cast_fail.wast.js-expected.txt

This is a testharness.js-based test.
Found 36 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] #1 Reinitialize the default imports
  promise_test: Unhandled rejection with value: object "TypeError: Cannot convert 10 to a BigInt"
[FAIL] #6 Test that WebAssembly compilation fails (br_on_cast_fail.wast:249)
  assert_equals: expected true but got false
[FAIL] #8 Test that WebAssembly compilation fails (br_on_cast_fail.wast:267)
  assert_equals: expected true but got false
[FAIL] #9 Test that WebAssembly compilation fails (br_on_cast_fail.wast:275)
  assert_equals: expected true but got false
[FAIL] #11 Reinitialize the default imports
  promise_test: Unhandled rejection with value: object "TypeError: Cannot convert 10 to a BigInt"
[FAIL] #12 Test that WebAssembly compilation succeeds (br_on_cast_fail.wast:3)
  assert_true: WebAssembly.compile failed unexpectedly with TypeError: Cannot convert 10 to a BigInt at {loc} expected true got false
[FAIL] #13 Test that WebAssembly instantiation succeeds
  assert_true: unexpected instantiation error, observed TypeError: Cannot read properties of undefined (reading 'source') \n    at instance (http://web-platform.test:8001/wasm/core/js/harness/async_index.js:207:15)\n    at br_on_cast_fail_wast_js (http://web-platform.test:8001/wasm/core/js/gc/br_on_cast_fail.wast.js:7:10)\n    at http://web-platform.test:8001/wasm/core/js/gc/br_on_cast_fail.wast.js:123:3 expected true got false
[FAIL] #14 run
  assert_true: unexpected runtime error, observed TypeError: Cannot read properties of undefined (reading 'init') \n    at run (http://web-platform.test:8001/wasm/core/js/harness/async_index.js:248:15)\n    at br_on_cast_fail_wast_js (http://web-platform.test:8001/wasm/core/js/gc/br_on_cast_fail.wast.js:10:1)\n    at http://web-platform.test:8001/wasm/core/js/gc/br_on_cast_fail.wast.js:123:3 expected true got false
[FAIL] #15 Test that a WebAssembly code returns a specific result (br_on_cast_fail.wast:71)
  assert_true: unexpected runtime error, observed TypeError: Cannot read properties of undefined (reading 'br_on_non_null') \n    at assert_return (http://web-platform.test:8001/wasm/core/js/harness/async_index.js:312:15)\n    at br_on_cast_fail_wast_js (http://web-platform.test:8001/wasm/core/js/gc/br_on_cast_fail.wast.js:13:1)\n    at http://web-platform.test:8001/wasm/core/js/gc/br_on_cast_fail.wast.js:123:3 expected true got false
[FAIL] #16 Test that a WebAssembly code returns a specific result (br_on_cast_fail.wast:72)
  assert_true: unexpected runtime error, observed TypeError: Cannot read properties of undefined (reading 'br_on_non_null') \n    at assert_return (http://web-platform.test:8001/wasm/core/js/harness/async_index.js:312:15)\n    at br_on_cast_fail_wast_js (http://web-platform.test:8001/wasm/core/js/gc/br_on_cast_fail.wast.js:16:1)\n    at http://web-platform.test:8001/wasm/core/js/gc/br_on_cast_fail.wast.js:123:3 expected true got false
[FAIL] #17 Test that a WebAssembly code returns a specific result (br_on_cast_fail.wast:73)
  assert_true: unexpected runtime error, observed TypeError: Cannot read properties of undefined (reading 'br_on_non_null') \n    at assert_return (http://web-platform.test:8001/wasm/core/js/harness/async_index.js:312:15)\n    at br_on_cast_fail_wast_js (http://web-platform.test:8001/wasm/core/js/gc/br_on_cast_fail.wast.js:19:1)\n    at http://web-platform.test:8001/wasm/core/js/gc/br_on_cast_fail.wast.js:123:3 expected true got false
[FAIL] #18 Test that a WebAssembly code returns a specific result (br_on_cast_fail.wast:74)
  assert_true: unexpected runtime error, observed TypeError: Cannot read properties of undefined (reading 'br_on_non_null') \n    at assert_return (http://web-platform.test:8001/wasm/core/js/harness/async_index.js:312:15)\n    at br_on_cast_fail_wast_js (http://web-platform.test:8001/wasm/core/js/gc/br_on_cast_fail.wast.js:22:1)\n    at http://web-platform.test:8001/wasm/core/js/gc/br_on_cast_fail.wast.js:123:3 expected true got false
[FAIL] #19 Test that a WebAssembly code returns a specific result (br_on_cast_fail.wast:75)
  assert_true: unexpected runtime error, observed TypeError: Cannot read properties of undefined (reading 'br_on_non_null') \n    at assert_return (http://web-platform.test:8001/wasm/core/js/harness/async_index.js:312:15)\n    at br_on_cast_fail_wast_js (http://web-platform.test:8001/wasm/core/js/gc/br_on_cast_fail.wast.js:25:1)\n    at http://web-platform.test:8001/wasm/core/js/gc/br_on_cast_fail.wast.js:123:3 expected true got false
[FAIL] #20 Test that a WebAssembly code returns a specific result (br_on_cast_fail.wast:77)
  assert_true: unexpected runtime error, observed TypeError: Cannot read properties of undefined (reading 'br_on_non_i31') \n    at assert_return (http://web-platform.test:8001/wasm/core/js/harness/async_index.js:312:15)\n    at br_on_cast_fail_wast_js (http://web-platform.test:8001/wasm/core/js/gc/br_on_cast_fail.wast.js:28:1)\n    at http://web-platform.test:8001/wasm/core/js/gc/br_on_cast_fail.wast.js:123:3 expected true got false
[FAIL] #21 Test that a WebAssembly code returns a specific result (br_on_cast_fail.wast:78)
  assert_true: unexpected runtime error, observed TypeError: Cannot read properties of undefined (reading 'br_on_non_i31') \n    at assert_return (http://web-platform.test:8001/wasm/core/js/harness/async_index.js:312:15)\n    at br_on_cast_fail_wast_js (http://web-platform.test:8001/wasm/core/js/gc/br_on_cast_fail.wast.js:31:1)\n    at http://web-platform.test:8001/wasm/core/js/gc/br_on_cast_fail.wast.js:123:3 expected true got false
[FAIL] #22 Test that a WebAssembly code returns a specific result (br_on_cast_fail.wast:79)
  assert_true: unexpected runtime error, observed TypeError: Cannot read properties of undefined (reading 'br_on_non_i31') \n    at assert_return (http://web-platform.test:8001/wasm/core/js/harness/async_index.js:312:15)\n    at br_on_cast_fail_wast_js (http://web-platform.test:8001/wasm/core/js/gc/br_on_cast_fail.wast.js:34:1)\n    at http://web-platform.test:8001/wasm/core/js/gc/br_on_cast_fail.wast.js:123:3 expected true got false
[FAIL] #23 Test that a WebAssembly code returns a specific result (br_on_cast_fail.wast:80)
  assert_true: unexpected runtime error, observed TypeError: Cannot read properties of undefined (reading 'br_on_non_i31') \n    at assert_return (http://web-platform.test:8001/wasm/core/js/harness/async_index.js:312:15)\n    at br_on_cast_fail_wast_js (http://web-platform.test:8001/wasm/core/js/gc/br_on_cast_fail.wast.js:37:1)\n    at http://web-platform.test:8001/wasm/core/js/gc/br_on_cast_fail.wast.js:123:3 expected true got false
[FAIL] #24 Test that a WebAssembly code returns a specific result (br_on_cast_fail.wast:81)
  assert_true: unexpected runtime error, observed TypeError: Cannot read properties of undefined (reading 'br_on_non_i31') \n    at assert_return (http://web-platform.test:8001/wasm/core/js/harness/async_index.js:312:15)\n    at br_on_cast_fail_wast_js (http://web-platform.test:8001/wasm/core/js/gc/br_on_cast_fail.wast.js:40:1)\n    at http://web-platform.test:8001/wasm/core/js/gc/br_on_cast_fail.wast.js:123:3 expected true got false
[FAIL] #25 Test that a WebAssembly code returns a specific result (br_on_cast_fail.wast:83)
  assert_true: unexpected runtime error, observed TypeError: Cannot read properties of undefined (reading 'br_on_non_struct') \n    at assert_return (http://web-platform.test:8001/wasm/core/js/harness/async_index.js:312:15)\n    at br_on_cast_fail_wast_js (http://web-platform.test:8001/wasm/core/js/gc/br_on_cast_fail.wast.js:43:1)\n    at http://web-platform.test:8001/wasm/core/js/gc/br_on_cast_fail.wast.js:123:3 expected true got false
[FAIL] #26 Test that a WebAssembly code returns a specific result (br_on_cast_fail.wast:84)
  assert_true: unexpected runtime error, observed TypeError: Cannot read properties of undefined (reading 'br_on_non_struct') \n    at assert_return (http://web-platform.test:8001/wasm/core/js/harness/async_index.js:312:15)\n    at br_on_cast_fail_wast_js (http://web-platform.test:8001/wasm/core/js/gc/br_on_cast_fail.wast.js:46:1)\n    at http://web-platform.test:8001/wasm/core/js/gc/br_on_cast_fail.wast.js:123:3 expected true got false
[FAIL] #27 Test that a WebAssembly code returns a specific result (br_on_cast_fail.wast:85)
  assert_true: unexpected runtime error, observed TypeError: Cannot read properties of undefined (reading 'br_on_non_struct') \n    at assert_return (http://web-platform.test:8001/wasm/core/js/harness/async_index.js:312:15)\n    at br_on_cast_fail_wast_js (http://web-platform.test:8001/wasm/core/js/gc/br_on_cast_fail.wast.js:49:1)\n    at http://web-platform.test:8001/wasm/core/js/gc/br_on_cast_fail.wast.js:123:3 expected true got false
[FAIL] #28 Test that a WebAssembly code returns a specific result (br_on_cast_fail.wast:86)
  assert_true: unexpected runtime error, observed TypeError: Cannot read properties of undefined (reading 'br_on_non_struct') \n    at assert_return (http://web-platform.test:8001/wasm/core/js/harness/async_index.js:312:15)\n    at br_on_cast_fail_wast_js (http://web-platform.test:8001/wasm/core/js/gc/br_on_cast_fail.wast.js:52:1)\n    at http://web-platform.test:8001/wasm/core/js/gc/br_on_cast_fail.wast.js:123:3 expected true got false
[FAIL] #29 Test that a WebAssembly code returns a specific result (br_on_cast_fail.wast:87)
  assert_true: unexpected runtime error, observed TypeError: Cannot read properties of undefined (reading 'br_on_non_struct') \n    at assert_return (http://web-platform.test:8001/wasm/core/js/harness/async_index.js:312:15)\n    at br_on_cast_fail_wast_js (http://web-platform.test:8001/wasm/core/js/gc/br_on_cast_fail.wast.js:55:1)\n    at http://web-platform.test:8001/wasm/core/js/gc/br_on_cast_fail.wast.js:123:3 expected true got false
[FAIL] #30 Test that a WebAssembly code returns a specific result (br_on_cast_fail.wast:89)
  assert_true: unexpected runtime error, observed TypeError: Cannot read properties of undefined (reading 'br_on_non_array') \n    at assert_return (http://web-platform.test:8001/wasm/core/js/harness/async_index.js:312:15)\n    at br_on_cast_fail_wast_js (http://web-platform.test:8001/wasm/core/js/gc/br_on_cast_fail.wast.js:58:1)\n    at http://web-platform.test:8001/wasm/core/js/gc/br_on_cast_fail.wast.js:123:3 expected true got false
[FAIL] #31 Test that a WebAssembly code returns a specific result (br_on_cast_fail.wast:90)
  assert_true: unexpected runtime error, observed TypeError: Cannot read properties of undefined (reading 'br_on_non_array') \n    at assert_return (http://web-platform.test:8001/wasm/core/js/harness/async_index.js:312:15)\n    at br_on_cast_fail_wast_js (http://web-platform.test:8001/wasm/core/js/gc/br_on_cast_fail.wast.js:61:1)\n    at http://web-platform.test:8001/wasm/core/js/gc/br_on_cast_fail.wast.js:123:3 expected true got false
[FAIL] #32 Test that a WebAssembly code returns a specific result (br_on_cast_fail.wast:91)
  assert_true: unexpected runtime error, observed TypeError: Cannot read properties of undefined (reading 'br_on_non_array') \n    at assert_return (http://web-platform.test:8001/wasm/core/js/harness/async_index.js:312:15)\n    at br_on_cast_fail_wast_js (http://web-platform.test:8001/wasm/core/js/gc/br_on_cast_fail.wast.js:64:1)\n    at http://web-platform.test:8001/wasm/core/js/gc/br_on_cast_fail.wast.js:123:3 expected true got false
[FAIL] #33 Test that a WebAssembly code returns a specific result (br_on_cast_fail.wast:92)
  assert_true: unexpected runtime error, observed TypeError: Cannot read properties of undefined (reading 'br_on_non_array') \n    at assert_return (http://web-platform.test:8001/wasm/core/js/harness/async_index.js:312:15)\n    at br_on_cast_fail_wast_js (http://web-platform.test:8001/wasm/core/js/gc/br_on_cast_fail.wast.js:67:1)\n    at http://web-platform.test:8001/wasm/core/js/gc/br_on_cast_fail.wast.js:123:3 expected true got false
[FAIL] #34 Test that a WebAssembly code returns a specific result (br_on_cast_fail.wast:93)
  assert_true: unexpected runtime error, observed TypeError: Cannot read properties of undefined (reading 'br_on_non_array') \n    at assert_return (http://web-platform.test:8001/wasm/core/js/harness/async_index.js:312:15)\n    at br_on_cast_fail_wast_js (http://web-platform.test:8001/wasm/core/js/gc/br_on_cast_fail.wast.js:70:1)\n    at http://web-platform.test:8001/wasm/core/js/gc/br_on_cast_fail.wast.js:123:3 expected true got false
[FAIL] #35 Test that a WebAssembly code returns a specific result (br_on_cast_fail.wast:95)
  assert_true: unexpected runtime error, observed TypeError: Cannot read properties of undefined (reading 'null-diff') \n    at assert_return (http://web-platform.test:8001/wasm/core/js/harness/async_index.js:312:15)\n    at br_on_cast_fail_wast_js (http://web-platform.test:8001/wasm/core/js/gc/br_on_cast_fail.wast.js:73:1)\n    at http://web-platform.test:8001/wasm/core/js/gc/br_on_cast_fail.wast.js:123:3 expected true got false
[FAIL] #36 Test that a WebAssembly code returns a specific result (br_on_cast_fail.wast:96)
  assert_true: unexpected runtime error, observed TypeError: Cannot read properties of undefined (reading 'null-diff') \n    at assert_return (http://web-platform.test:8001/wasm/core/js/harness/async_index.js:312:15)\n    at br_on_cast_fail_wast_js (http://web-platform.test:8001/wasm/core/js/gc/br_on_cast_fail.wast.js:76:1)\n    at http://web-platform.test:8001/wasm/core/js/gc/br_on_cast_fail.wast.js:123:3 expected true got false
[FAIL] #37 Test that a WebAssembly code returns a specific result (br_on_cast_fail.wast:97)
  assert_true: unexpected runtime error, observed TypeError: Cannot read properties of undefined (reading 'null-diff') \n    at assert_return (http://web-platform.test:8001/wasm/core/js/harness/async_index.js:312:15)\n    at br_on_cast_fail_wast_js (http://web-platform.test:8001/wasm/core/js/gc/br_on_cast_fail.wast.js:79:1)\n    at http://web-platform.test:8001/wasm/core/js/gc/br_on_cast_fail.wast.js:123:3 expected true got false
[FAIL] #38 Test that a WebAssembly code returns a specific result (br_on_cast_fail.wast:98)
  assert_true: unexpected runtime error, observed TypeError: Cannot read properties of undefined (reading 'null-diff') \n    at assert_return (http://web-platform.test:8001/wasm/core/js/harness/async_index.js:312:15)\n    at br_on_cast_fail_wast_js (http://web-platform.test:8001/wasm/core/js/gc/br_on_cast_fail.wast.js:82:1)\n    at http://web-platform.test:8001/wasm/core/js/gc/br_on_cast_fail.wast.js:123:3 expected true got false
[FAIL] #39 Test that a WebAssembly code returns a specific result (br_on_cast_fail.wast:99)
  assert_true: unexpected runtime error, observed TypeError: Cannot read properties of undefined (reading 'null-diff') \n    at assert_return (http://web-platform.test:8001/wasm/core/js/harness/async_index.js:312:15)\n    at br_on_cast_fail_wast_js (http://web-platform.test:8001/wasm/core/js/gc/br_on_cast_fail.wast.js:85:1)\n    at http://web-platform.test:8001/wasm/core/js/gc/br_on_cast_fail.wast.js:123:3 expected true got false
[FAIL] #47 Test that WebAssembly compilation fails (br_on_cast_fail.wast:249)
  assert_true: \n    at module (http://web-platform.test:8001/wasm/core/js/harness/async_index.js:163:15)\n    at assert_invalid (http://web-platform.test:8001/wasm/core/js/harness/async_index.js:192:3)\n    at br_on_cast_fail_wast_js (http://web-platform.test:8001/wasm/core/js/gc/br_on_cast_fail.wast.js:109:1)\n    at http://web-platform.test:8001/wasm/core/js/gc/br_on_cast_fail.wast.js:123:3 expected true got false
[FAIL] #49 Test that WebAssembly compilation fails (br_on_cast_fail.wast:267)
  assert_true: \n    at module (http://web-platform.test:8001/wasm/core/js/harness/async_index.js:163:15)\n    at assert_invalid (http://web-platform.test:8001/wasm/core/js/harness/async_index.js:192:3)\n    at br_on_cast_fail_wast_js (http://web-platform.test:8001/wasm/core/js/gc/br_on_cast_fail.wast.js:115:1)\n    at http://web-platform.test:8001/wasm/core/js/gc/br_on_cast_fail.wast.js:123:3 expected true got false
[FAIL] #50 Test that WebAssembly compilation fails (br_on_cast_fail.wast:275)
  assert_true: \n    at module (http://web-platform.test:8001/wasm/core/js/harness/async_index.js:163:15)\n    at assert_invalid (http://web-platform.test:8001/wasm/core/js/harness/async_index.js:192:3)\n    at br_on_cast_fail_wast_js (http://web-platform.test:8001/wasm/core/js/gc/br_on_cast_fail.wast.js:118:1)\n    at http://web-platform.test:8001/wasm/core/js/gc/br_on_cast_fail.wast.js:123:3 expected true got false
Harness: the test ran to completion.
